requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.349
langchain-google-genai==0.0.11
google-generativeai==0.3.2
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0